backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
gTTS==2.3.2
playsound==1.3.0
torch>=2.0.0
transformers>=4.30.0
opencv-python-headless>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
requests>=2.28.0
python-dotenv>=1.0.0
